Latest Version 00:22 08-11-2024

+++ School Connect/SchoolConnect_RepositoryLayer/Interfaces/ILearner.cs
@@ -11,7 +11,7 @@ namespace SchoolConnect_RepositoryLayer.Interfaces
         Task<Dictionary<string, object>> GetLearnerByIdNoAsync(string learnerIdNo);
         Task<Dictionary<string, object>> GetAll();
         Task<Dictionary<string, object>> GetLearnersByMainTeacherAsync(long teacherId);
-        Task<Dictionary<string, object>> GetLearnersByClassID(long classId);
+        Task<Dictionary<string, object>> GetLearnersByClassID(int classId);
         Task<Dictionary<string, object>> UpdateAsync(Learner learner);
         Task<Dictionary<string, object>> Remove(long learnerId);
         Task<Dictionary<string, object>> RemoveClass(string classId);
